Cleaned up auth code, please refer to issue #32
I’ve made it so the admin user doesn’t need to be logged in, and being
logged in isn’t enough, they have to have the token set (which gets set
when they setup Plex Auth), added new setting variable which is the
admin’s Plex username to allow the above. Also once we tagged movies
with the requesting user the admin user gets tagged as well.

Cleaned up the home.html for when the user doesn’t have a token so it
fits more in style with the rest of the application and to give them
some info as to what they’re entering their username for. Will do the
same for the Plex page, will make it similar to the Couch Potato page.

Added alert dialogs for when the user isn't on the friends list and for
when admin user doesn't authenticate against Plex.tv (the Plex.tv one
is rudimentary and doesn't check if it's a 401 or anything, just takes
anything that isn't a 201 and complains).

Moved some code around and cleaned up in a few small places, nothing
major just so I could read it easier.