Updated plex auth
Plex auth now requires logging in to your Plex.tv account with username
and password. This is to improve security but is better than making
users create a new login just for this app.

Also improved the router auth. Even if there was an auth key saved, it
didn’t have to be set to true to allow searching and requesting.