Add pushbullet channels; Add auto-migrate database feature
## Pushbullet Channels
This update adds the ability to send notifications to custom made
Pushbullet channels which can be subscribed to by your PlexRequest users
so they can receive notifications ftom the service.

Checkout the README for more info on creating a custom pushbullet channel.

## Auto-migration of DB collections

While not a sexy update on tge surface, it addresses a core issue with
 this application that has hindered its forward progress.

The main purpose of this change is to allow the mongodb databases currently in use to to continue being used after an update to the applications code makes a change to a collections schema. Before this update, users had no easy way to upgrade their applications without loosing the data that had already been collecting.

Cue the davidyaha:collection2-migrations! This package monitors each
collections schema and will make the necessary adjustments to the records
in the affected collection on the fly.

All this means is that developers will now be able to add new features without being bound to the current versions database design.