Overhaul of Logging System - Permanently fixin-app version checking. - Create a wrapper around the Meteor 'Check()' function to allow better logging of what function arguments are not passing their checks. - Refactor and simplify Sonarr and Radarr's methods and structure. - Begin removing redundant code from the service helpers and consolodating it into their classes where possible. - Did I mention the logging overhaul?
Added a TODO file in the project root directory that outlines what is
still left to finish from this current sprint.

I know the authentication issue is still there and I haven't forgotten
about it. The logging issue is something that has been needed for a
while and the new check function wrapper will hopefully make handling
the authentication logic much easier.

Signed-off-by: Ricky Grassmuck <rigrassm@gmail.com>